+++ depth_player/video/reader.h
@@ -51,7 +51,7 @@ public:
 
 };
 
-AVFrame *alloc_audio_frame(enum AVSampleFormat sample_fmt, uint64_t channel_layout, int sample_rate, int nb_samples);
+AVFrame *alloc_audio_frame(enum AVSampleFormat sample_fmt, uint64_t channel_layout, int sample_rate, int64_t nb_samples);
 AVFrame *alloc_picture(enum AVPixelFormat pix_fmt, int width, int height);
 
 struct InputStream {
